Computing new_Y before new_X, early abort speedup.
By re-orienting the points data in the precalculation
done in main into a [X, Z, Y] form, we can compute
things in this order, which allows for early abort:

    wxnew = old_X-mcos
    y = 96 - (old_Z/wxnew)
    if y is bad, early abort <==== HERE
    x = 128 + ((old_Y+msin)/wxnew)

This gives another speed boost. We are now at 9.9 fps,
so close to the barrier of 10! :-)

Also: much better documentation of what happens on
the stack during the computations of wxnew, new_Y and new_X.